+++ CHANGES.rst
@@ -25,6 +25,7 @@ Added
 
 - Python 3.11 support. :pr:`61`
 - apparmor slapd configuration instructions in CONTRIBUTING.rst :pr:`66`
+- ``preferredLanguage`` attribute support. :pr:`75`
 
 Changed
 *******